Refactored ImageReader to move endianess related methods to EndianImageReader
