+++ src/Core/Core.csproj
@@ -199,6 +199,7 @@
     <Compile Include="Configuration\UiPreferencesConfiguration.cs" />
     <Compile Include="Configuration\UiStyleElement.cs" />
     <Compile Include="Endianess.cs" />
+    <Compile Include="EndianImageReader.cs" />
     <Compile Include="EnumerableEx.cs" />
     <Compile Include="EvaluationContext.cs" />
     <Compile Include="EventEx.cs" />
