+++ src/Core/ImageLoader.cs
@@ -29,7 +29,7 @@ namespace Reko.Core
     /// generate a Program after carrying out relocations, resolving external
     /// symbols etc.
 	/// </summary>
-	public abstract class ImageLoader
+	public class ImageLoader
 	{
         public ImageLoader(IServiceProvider services, string filename, byte[] imgRaw)
         {
@@ -39,11 +39,6 @@ namespace Reko.Core
         }
 
         public IServiceProvider Services { get; private set; }
-
-        /// <summary>
-        /// If nothing else is specified, this is the address at which the image will be loaded.
-        /// </summary>
-        public abstract Address PreferredBaseAddress { get; set; }
 
         /// <summary>
         /// Optional loader-specific argument specified in app.config.
